Reliably disable body scroll when modal is open
iOS Safari doesn’t behave like other browsers when it comes to
disabling scrolling. Simply setting overflow: hidden on the body
element works in every other browser, but only setting the body to
position: fixed with overflow: hidden works in iOS Safari. However,
that resets the user’s scroll position to the top of the page. The best
solution I could find was to offset the body at the same time so the
content is in the same position relative to the window. Again, this
didn’t work in iOS Safari, which seems to render the modal relative to
the body, so the modal would be off the screen. Instead, I’m wrapping
all the body content in a top-level div and applying the offset to that
div.

Disabling body scroll on iOS Safari also fixes their bug that causes
forms in modals (and other fixed-position elements) to become unusable.