Fix issue with Chrome for iOS
Chrome for iOS wasn’t rendering the image fast enough to be able to
calculate the height in time to fit the modal to the screen. Appending
the preloaded image object (instead of appending a new image object
with the same source, which was inefficient) and setting the width and
height attributes fix the problem and make this more robust.